refactor(App): extract useHealth composable
Groups geminiAvailable + sandboxEnabled + fetchHealth into a single
self-contained composable, narrowing their scope and removing 15
lines from App.vue.
